BOCM-80 moves getValidation controller tests to own file
